код поправлен согласно полученным комментариям
